Moving activities to the correct locations
