+++ LanguageRecorder/app/src/main/res/layout/activity_village.xml
@@ -4,7 +4,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.example.lachlan.myfirstapp.code.VillageActivity">
+    tools:context="com.example.lachlan.myfirstapp.VillageActivity">
 
     <Spinner
         android:layout_width="wrap_content"
