+++ LanguageRecorder/app/src/main/res/menu/menu_sub_district.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.example.lachlan.myfirstapp.code.SubDistrictActivity">
+    tools:context="com.example.lachlan.myfirstapp.SubDistrictActivity">
     <item android:id="@+id/action_settings" android:title="@string/action_settings"
         android:orderInCategory="100" app:showAsAction="never" />
 </menu>
